requirements.txt
Django==5.1
djangorestframework==3.15.2
google-cloud-storage==2.18.2
pillow==10.4.0
python-dotenv==1.0.1
